gRPC Core Client interceptor wrong status code and description if writing to a client or duplex streaming service fails (#132)
Any exceptions during a write to a clientstreaming or duplexstreaming RPC are not tracked. They are bubbling up to the caller and we are losing the opportunity to record the correct status code and properly stop the activity. Before this change, the newly added tests would fail with a status code of (1) cancelled which was getting set during Dispose instead of proactively at the time of the failure.