Update TFM ordering for VS Code / Omnisharp
OmniSharp requires the set of library TFMs to be in descending order:

- .NET core versions
- .NET standard versions
- .NET desktop versions

It uses the first one in the list as the source for analyzers and falls
back accordingly.

Updated projects with multiple TFMs to correct the order and improve
analysis support for VS Code/OmniSharp, including a note to help
maintain that order.