chore: enable PublicApiAnalyzers on MacOS X
The PublicApiAnalyzers is currently set to run only on Windows. I suspect it is so that, during the build process, one would have the Linux builds proceed and provide better insight to other types of failures.

But the current setting leaves people using a Mac out in the cold. This commit switches up the condition to be `$(OS) != 'Linux'`, reducing the blast radius.

(In reality, a better fix would be to skip these checks if the build process is running in CI, as they also occur during the Code Coverage step. Or maybe they should be a step on its own?)