Remove targets for now
For now, we just want to let this PR focus on the detector and
introducing a convention and mechanism for making build-time
resource attributes available for detection at runtime by the
assembly metadata resource detector.

The `vcs.*` attrs may return in a forthcoming PR, but TBD for now.
In the meantime, a draft is available for reference at:
- austindrenski/opentelemetry-dotnet-contrib#1