fix(OpenTelemetry.Resources.AWS.Tests): Incorrect unit test result when running on EC2
When the unit test in AWSEC2DetectorTests runs on an EC2 machine, it reports a failure.
The test assumes it will never run on an EC2 instance, which is not a safe assumption.

This fix adds a conditional check to detect if the current environment is EC2
and adjusts the assertion accordingly.

dotnet build: pass
dotnet test: pass