Add `HttpClientFactory` setting to `OpAmpClientSettings`
Introduce a new `HttpClientFactory` setting in the `OpAmpClientSettings` class to allow customization of `HttpClient` instances for OpAMP HTTP transport. Update `PlainHttpTransport` to utilize this factory, removing the `HttpClientHandler`. Modify `OpAmpClient` to pass settings to `PlainHttpTransport`.

Add a test to verify the use of the configured `HttpClientFactory`, including custom headers. Update `OpAmpFakeHttpServer` to capture HTTP request headers and provide a method to retrieve them for testing.

Update `CHANGELOG.md` and `PublicAPI.Unshipped.txt` to document the new setting.